% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getnetstats.R
\name{getnetstats}
\alias{getnetstats}
\title{Get micro-level network measures for words in the single layer language network.}
\usage{
getnetstats(network)
}
\arguments{
\item{network}{An igraph object of the language network.}
}
\value{
A dataframe containing network statistics for each node in the network.
}
\description{
Get micro-level network measures for words in the single layer language network.
}
\examples{
data <- getnetstats(network)
head(data)
}
