% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getmultistats.R
\name{getmultistats}
\alias{getmultistats}
\title{Get micro-level network statistics for words in the language network.}
\usage{
getmultistats(network)
}
\arguments{
\item{network}{An igraph object of the language network.}
}
\value{
A dataframe containing network statistics for each node in the network.
}
\description{
Get micro-level network statistics for words in the language network.
}
\examples{
data <- getnetstats(network)
head(data)
}
