% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makelangnet.R
\name{makelangnet}
\alias{makelangnet}
\title{Converts a list of words to an igraph network object. Note that makelangnet is simply a wrapper for tolangnet and nodeindex}
\usage{
makelangnet(wordlist)
}
\arguments{
\item{wordlist}{A list of words. Must be a character vector.}
}
\value{
An igraph object of the language network created from \code{wordlist}.
}
\description{
Converts a list of words to an igraph network object. Note that makelangnet is simply a wrapper for tolangnet and nodeindex
}
\examples{
somewords <- c('cat', 'bat', 'cap', 'cape')
somewordsnet <- tolangnet(somewords)
plot(somewordsnet) #plots the graph
}
